# dfras-docs/dfras-api-swagger.yaml
openapi: 3.0.3
info:
  title: DFRAS API
  description: |
    Delivery Failure Root Cause Analysis System (DFRAS) API
    
    A comprehensive AI-powered platform for analyzing delivery failures and providing predictive insights for logistics operations.
    
    ## Key Features
    - Multi-source data integration (orders, fleet logs, warehouse records, customer feedback)
    - AI-powered root cause analysis using LLM (all-MiniLM-L6-v2)
    - Automated correlation engine for pattern recognition
    - Intelligent reporting with actionable recommendations
    - Predictive analytics for proactive failure prevention
    
    ## Authentication
    All endpoints require JWT authentication. Use the `/auth/login` endpoint to obtain a token.
    
    ## Sample Data
    The system includes comprehensive sample data from `third-assignment-sample-data-set`:
    - 750+ clients across multiple cities
    - 15,000+ orders with delivery status
    - 50+ warehouses with capacity data
    - 200+ drivers with performance metrics
    - External factors (weather, traffic, events)
    - Customer feedback and ratings
  version: 1.0.0
  contact:
    name: DFRAS Support
    email: support@dfras.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local Development Server
  - url: http://localhost:8002
    description: Analytics Service Direct Access
  - url: http://localhost:8001
    description: Data Service Direct Access
security:
  - bearerAuth: []
paths:
...
  /api/ai/analyze:
    post:
      tags:
        - AI Query
      summary: Analyze Query
      description: Process natural language queries using AI for delivery failure analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Natural language query about delivery failures
                  example: "Why were deliveries delayed in Mumbai yesterday?"
                context:
                  type: string
                  description: Analysis context
                  example: "delivery_delay_analysis"
                filters:
                  type: object
                  description: Additional filters for the analysis
                  properties:
                    city:
                      type: string
                      example: "Mumbai"
                    date:
                      type: string
                      example: "yesterday"
                    status:
                      type: string
                      example: "delayed"
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                    example: "Why were deliveries delayed in Mumbai yesterday?"
                  analysis:
                    type: string
                    example: "Based on the analysis of delivery data for Mumbai yesterday, the main causes of delays were..."
                  insights:
                    type: array
                    items:
                      type: object
                      properties:
                        insight_type:
                          type: string
                          example: "weather_impact"
                        description:
                          type: string
                          example: "Heavy rainfall in Mumbai caused 35% of delivery delays"
                        impact_score:
                          type: number
                          format: float
                          example: 0.85
                        recommendations:
                          type: array
                          items:
                            type: string
                          example: ["Implement weather-based routing", "Increase buffer time during monsoon"]
                  data_summary:
                    type: object
                    properties:
                      total_deliveries:
                        type: integer
                        example: 500
                      delayed_deliveries:
                        type: integer
                        example: 75
                      delay_rate:
                        type: number
                        format: float
                        example: 15.0
                  confidence_score:
                    type: number
                    format: float
                    example: 0.92
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid query or parameters
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...